+++ example/android/app/src/main/kotlin/com/example/net_checker_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.net_checker_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()