+++ lib/src/core/latency_checker_io.dart
@@ -3,15 +3,29 @@ import '../config/internet_config.dart';
 
 /// IO implementation of latency checking.
 ///
-/// This runs on:
-/// Android, iOS, Windows, macOS, Linux
+/// Supported platforms:
+/// - Android
+/// - iOS
+/// - Windows
+/// - macOS
+/// - Linux
+///
+/// Uses DNS lookup to measure latency.
 class LatencyCheckerImpl {
   /// Measures latency using DNS lookup.
+  ///
+  /// Returns:
+  /// - latency in milliseconds
+  /// - `null` if failed
   static Future<int?> getLatency(InternetConfig config) async {
+    final uri = config.parsedLatencyHost;
+
+    if (uri == null || uri.host.isEmpty) return null;
+
     final stopwatch = Stopwatch()..start();
 
     try {
-      await InternetAddress.lookup(config.latencyHost);
+      await InternetAddress.lookup(uri.host).timeout(config.timeout);
 
       stopwatch.stop();